# Support new OCI image index content type to get docker image digest from registry (content digest)