- ubuntu:20.04 is now only available with new 'application/vnd.oci.image.index.v1+json'
- ubuntu:12.04 is only available with old 'application/vnd.docker.distribution.manifest.v2+json' (and should hopefully stay that way in the future, for tests stability)
- asking only the new oci.image.index triggers an implicit fallback to 'application/vnd.docker.distribution.manifest.v1+prettyjws' for old images, with a different digest, we don't want that (we want digest stability for base image cache)
- `Docker-Content-Digest` header is still a valid digest with the new content type

=> ask both content types, priority to new one.

Added a quick unit test testing get_image_digest on both old and new content-types (hopefully it should stay stable on docker hub side; if it changes, we should adapt it).

Manually tested test failures when removing either of the accepted_content_types.