@thomas-riccardi 2023-02-01
Related:
- https://www.docker.com/blog/announcing-docker-hub-oci-artifacts-support/
-  https://github.com/containrrr/watchtower/issues/1528
- https://stackoverflow.com/questions/62267417/difference-between-oci-image-manifest-and-docker-v2-2-image-manifest